# infra/api-spec.yaml
openapi: 3.0.1
info:
  title: Chat API
  description: Conversational AI API with DynamoDB memory and Bedrock integration
  version: 1.0.0

paths:
  /chat:
    options:
      summary: CORS support
      description: Enable CORS for browsers
      responses:
        "200":
          description: Default CORS response
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: '{"statusCode": 200}'
        responses:
          default:
            statusCode: 200
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'"
              method.response.header.Access-Control-Allow-Methods: "'OPTIONS,GET,POST'"
              method.response.header.Access-Control-Allow-Origin: "'*'"

    post:
      summary: Send a chat message
      description: Send a user message and trigger async processing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - session_id
                - content
              properties:
                user_id:
                  type: string
                  example: "123"
                session_id:
                  type: string
                  example: "abc123"
                content:
                  type: string
                  example: "What's the weather today?"
      responses:
        "202":
          description: Message accepted for processing
          content:
            application/json:
              schema:
                type: object
                properties:
                  message_id:
                    type: string
                  status:
                    type: string
                  timestamp:
                    type: string
        "400":
          description: Bad request
        "500":
          description: Internal server error
      x-amazon-apigateway-integration:
        uri: ${lambda_invoke_arn}
        httpMethod: POST
        type: aws_proxy
        passthroughBehavior: when_no_match

  /messages:
    options:
      summary: CORS support
      description: Enable CORS for browsers
      responses:
        "200":
          description: Default CORS response
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: '{"statusCode": 200}'
        responses:
          default:
            statusCode: 200
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'"
              method.response.header.Access-Control-Allow-Methods: "'OPTIONS,GET,POST'"
              method.response.header.Access-Control-Allow-Origin: "'*'"

    get:
      summary: Get conversation messages
      description: Retrieve messages for a specific session
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
          example: "123"
        - name: session_id
          in: query
          required: true
          schema:
            type: string
          example: "abc123"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
          example: 50
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      type: object
                  count:
                    type: integer
        "400":
          description: Bad request
        "500":
          description: Internal server error
      x-amazon-apigateway-integration:
        uri: ${lambda_invoke_arn}
        httpMethod: POST
        type: aws_proxy
        passthroughBehavior: when_no_match

  /sessions/{user_id}:
    options:
      summary: CORS support
      description: Enable CORS for browsers
      responses:
        "200":
          description: Default CORS response
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: '{"statusCode": 200}'
        responses:
          default:
            statusCode: 200
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'"
              method.response.header.Access-Control-Allow-Methods: "'OPTIONS,GET,POST'"
              method.response.header.Access-Control-Allow-Origin: "'*'"

    get:
      summary: Get user sessions
      description: Retrieve all active sessions for a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          example: "123"
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  sessions:
                    type: array
                    items:
                      type: string
        "500":
          description: Internal server error
      x-amazon-apigateway-integration:
        uri: ${lambda_invoke_arn}
        httpMethod: POST
        type: aws_proxy
        passthroughBehavior: when_no_match

  /health:
    options:
      summary: CORS support
      description: Enable CORS for browsers
      responses:
        "200":
          description: Default CORS response
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: '{"statusCode": 200}'
        responses:
          default:
            statusCode: 200
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'"
              method.response.header.Access-Control-Allow-Methods: "'OPTIONS,GET,POST'"
              method.response.header.Access-Control-Allow-Origin: "'*'"

    get:
      summary: Health check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  service:
                    type: string
      x-amazon-apigateway-integration:
        uri: ${lambda_invoke_arn}
        httpMethod: POST
        type: aws_proxy
        passthroughBehavior: when_no_match